address gregs feedback
